back/auth.py
from jose import jwt
from passlib.context import CryptContext

def compare_passwords(plain_password, hashed_password):
    pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
    return pwd_context.verify(plain_password, hashed_password)