back/pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = ""
authors = ["Petar Vuković <petarvukovic50@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.31.1"
fastapi = "^0.115.0"
supabase = "^2.9.0"
load-dotenv = "^0.1.0"
bcrypt = "^4.2.0"
pydantic = {extras = ["email"], version = "^2.9.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
